"Short" casts

+++ src/pic/simulator/parser/commands/Andlw.java
@@ -31,7 +31,7 @@ public class Andlw extends Command
 	@Override
 	public void execute(PicProcessor proc) 
 	{
-		byte val = (byte) (proc.workRegister & arg0);
+		short val = (short) (proc.workRegister & arg0);
 		proc.workRegister = val;
 		
 		affectZeroBit(proc, val);
