+++ src/pic/simulator/specialfunctionregisters/Eedata.java
@@ -4,7 +4,7 @@ import pic.simulator.SpecialFunctionRegister;
 
 public class Eedata extends SpecialFunctionRegister
 {
-    private byte value;
+    private short value;
 
     public Eedata()
     {
@@ -12,13 +12,13 @@ public class Eedata extends SpecialFunctionRegister
     }
 
     @Override
-    public void setValue(byte value)
+    public void setValue(short value)
     {
         this.value = value;
     }
 
     @Override
-    public byte getValue()
+    public short getValue()
     {
         return value;
     }
